Create shipping.py
This is a project in the Codecademy Python3 Course. Sal's Shipping is a project in the flow control section that has you create a calculator for shipping rates. I went above by adding a sort feature that arranges the options from cheapest to most expensive, although to be fair the AI Learning Assistant provided me the support to needed to be able to do that.