Create Gradebook.py
This is a project in the Codecademy Python3 Course. Gradebook is a project in Codecademy's Python 3 course Lists module. In this project I created and modified a verity of lists using a few different methods. one method was using .append and .remove to add and remove data. The other was by setting one of the variables within a list to add 5 to itself. At the end I combined two separate 2D lists into a single 2D list forming the final grade book. I then added to the project by appending ["Python, 100"] and printed it to the terminal.