Create Len'sSlice
In this project I used a variety of methods and functions to manipulate both 2d and 3d lists. I used the .zip function to combine two lists into one 2d list. Once the list was combine I sorted it to list the slices from lowest to highest cost. From here we had some orders requiring the addition and removal of some items from the menu. At the end of the project I sorted the three cheapest Slices for three mice that wasted the cheapest slices and printed it to the screen. The mice were kind enough to add a 5star review so I added a review section to the project as well.