Thread Shed
This project in CodeCademy had me working with a very long string of data. Starting from a big list of every transaction I split and sorted the data into a usable format allowing for me to perform data analytics on the data. In the end I was able to break the string down far enough to get a count of each thread sold by color.