Create TimeTravelersToolkit.py
In this project I created a tool for time travelers that will output the details of their trip. It generates a random day between the years 1800 and 2100 then calculates a cost using the difference between the day traveled from and the day traveled to * 4. This tool also will generate a random continent between Europe, North America, and Africa for you to travel to.