Create hacked_fender.py
This lab had me working with different file formats including txt, json and csv. In the first part I grabbed all of the compromised usernames from the data leak file and then saved them to a file named compromised_users.txt. The next step was to create a json file containing a success message to send on to the boss. After that I created a new password file to replace the data leak with so that the hacker wouldn't have our precious passwords any longer.